requirements.txt
Flask==2.3.2
pandas==2.0.3
requests==2.28.2
PyYAML==6.0.1
google-cloud-secret-manager==5.6.0
google-cloud-spanner==7.11.0
firebase-admin==12.3.0